feat: expose basic non-topological sorted graph via from_d8_raw and from_cama_nextxy_raw
