+++ src/earthkit/hydro/_readers/__init__.py
@@ -1,8 +1,4 @@
-from .readers import (
-    find_main_var,
-    from_cama_downxy,
-    from_cama_nextxy,
-    from_d8,
-    from_grit,
-    import_earthkit_or_prompt_install,
-)
+from ._cama import from_cama_downxy, from_cama_nextxy
+from ._core import assign_coords, find_main_var, import_earthkit_or_prompt_install
+from ._d8 import from_d8
+from ._grit import from_grit
